requirements.txt
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==2.1.7
django-braces==1.14.0
django-oauth-toolkit==1.3.2
django-phonenumber-field==2.4.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.2
idna==2.9
oauthlib==3.1.0
phonenumbers==8.10.11
pkg-resources==0.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
urllib3==1.25.9